% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHfunctions.R
\name{plot_trends_over_time}
\alias{plot_trends_over_time}
\title{Visualize trends over time}
\usage{
plot_trends_over_time(
  data,
  x_var,
  y_vars,
  title = NULL,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{data}{A data frame containing the dataset to visualize. The first column should represent the time variable, and subsequent columns should represent different variables to plot.}

\item{x_var}{The name of the column representing the time variable.}

\item{y_vars}{A character vector containing the names of the columns representing the variables to plot.}

\item{title}{The title of the plot (optional).}

\item{xlab}{Label for the x-axis (optional).}

\item{ylab}{Label for the y-axis (optional).}
}
\value{
A line plot visualizing trends over time.
}
\description{
This function creates a line plot to visualize trends over time of the provided dataset.
}
