% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHMLfunctions.R
\name{predict_classifier}
\alias{predict_classifier}
\title{Make predictions using a classifier model}
\usage{
predict_classifier(model, new_data)
}
\arguments{
\item{model}{A trained classifier object}

\item{new_data}{A data frame containing new or unseen data to make predictions on}
}
\value{
a vector of predicted class labels
}
\description{
Make predictions using a classifier model
}
\examples{
\dontrun{
predictions <- predict_classifier(model= logistic_model, new_data = new_health<data)
}
}
