% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHMLfunctions.R
\name{train_classifier}
\alias{train_classifier}
\title{Train a classifier model with patient data}
\usage{
train_classifier(data, method)
}
\arguments{
\item{data}{a data frame with labeled patient data with features as columns and the target variable as a factor column}

\item{method}{the classification method to use (e.g. "logistic_regression","decision_tree","rain_forest")}
}
\value{
a trained classifier model object
}
\description{
Train a classifier model with patient data
}
\examples{
\dontrun{
model <- train_classifier(data = patient_data, method = "logistic_regression")
}
}
